requirements.txt
Django==1.9.2
Pillow==2.9.0
django-import-export==0.4.5
